requirements.txt

### **requirements.txt**

```txt
PyQt5==5.15.7
pyqrcode==1.2.1
Pillow==9.3.0
